requirements.txt
Flask==2.1.1
Pillow
webcolors==1.12
